// src/app/models/reparacionEnc.ts
export class reparacionEnc {
    idreparacionenc: number;
    nombrecliente: string;
    telefonocliente: string;
    duicliente: string;
    descripcion: string;
    recibo: string;
    idsucursal: number;
    estado: boolean;
    iduseradd: number;
    dateadd: string;
    idusermod: number;
    datemod: string;
}