// src/app/models/reparacionDet.ts
export class reparacionDet {
    idreparaciondet: number;
    idreparacionenc: number;
    nombreproducto: string;
    idproducto: number;
    cantidad: number;
    nota: string;
    categoriaproducto: string;
    precioregular: number;
    preciototal: number;
    iduseradd: number;
    dateadd: string;
    idusermod: number;
    datemod: string;
}